auth!

+++ public/app.js
@@ -2151,10 +2151,7 @@ function addContent(isAdmin) {
       showmodal.classList.remove("is-active");
     });
   } else {
-    let points_content = `
-    <div class="welcome is-size-3 has-text-centered">
-    </div>
-    <div class="columns is-centered mt-4">
+    let points_content = `<div class="columns is-centered mt-4">
     <div class="column pr-outer">
         <div class="card px-4 py-3 has-text-centered" event-type="volunteer">
             <header class="card-header">
@@ -2256,8 +2253,6 @@ function addContent(isAdmin) {
       .then((usersSnapshot) => {
         if (!usersSnapshot.empty) {
           let userDoc = usersSnapshot.docs[0];
-          let userData = userDoc.data();
-          let fullName = userData.full_name;
 
           userDoc.ref
             .collection("member_points")
@@ -2271,13 +2266,6 @@ function addContent(isAdmin) {
                   memberTotalPoints[eventType] += points;
                 }
               });
-              const firstPointDoc = pointsSnapshot.docs[0];
-              const semester = firstPointDoc
-                ? firstPointDoc.get("pointSemester").toLowerCase()
-                : "Unknown";
-
-              let welcomeDiv = document.querySelector(".welcome");
-              welcomeDiv.innerHTML = `Welcome ${fullName}! Here is your point summary for the ${semester} semester.`;
 
               // After all data is aggregated, update the UI
               updateCardsWithPoints(memberTotalPoints);
@@ -2573,29 +2561,33 @@ r_e("blog-link").addEventListener("click", () => {
             const post = doc.data();
             const postId = doc.id;
             let editDeleteHtml = "";
-            if (auth.currentUser.email === "amauwmadison@gmail.com") {
+            if (
+              auth.currentUser &&
+              auth.currentUser.email === "amauwmadison@gmail.com"
+            ) {
+              // User is authenticated and is "ama exec"
               editDeleteHtml = `
-                          <footer class="card-footer">
-                              <a href="#" class="card-footer-item" onclick="editPost('${postId}')">Edit</a>
-                              <a href="#" class="card-footer-item" onclick="deletePost('${postId}')">Delete</a>
-                          </footer>`;
+                <footer class="card-footer">
+                  <a href="#" class="card-footer-item" onclick="editPost('${postId}')">Edit</a>
+                  <a href="#" class="card-footer-item" onclick="deletePost('${postId}')">Delete</a>
+                </footer>`;
             }
             html += `
-                      <div class="container my-4">
-                          <div class="card" id="${postId}">
-                              <header class="card-header">
-                                  <p class="card-header-title">${post.title}</p>
-                              </header>
-                              <div class="card-content">
-                                  <div class="content">
-                                      ${post.message}
-                                      <br><br>
-                                      <p class="card-header-subtitle"><span style="font-size: smaller; font-weight: bold;">By: ${post.author} // <time datetime="${post.date}">${post.date}</time></span></p>
-                                  </div>
-                              </div>
-                              ${editDeleteHtml}
-                          </div>
-                      </div>`;
+              <div class="container my-4">
+                <div class="card" id="${postId}">
+                  <header class="card-header">
+                    <p class="card-header-title">${post.title}</p>
+                  </header>
+                  <div class="card-content">
+                    <div class="content">
+                      ${post.message}
+                      <br><br>
+                      <p class="card-header-subtitle"><span style="font-size: smaller; font-weight: bold;">By: ${post.author} // <time datetime="${post.date}">${post.date}</time></span></p>
+                    </div>
+                  </div>
+                  ${editDeleteHtml}
+                </div>
+              </div>`;
           });
           document.querySelector("#all_posts").innerHTML = html;
         })
